validation + logging

+++ src/idea/idea/idea.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, Logger } from '@nestjs/common';
+import { Injectable, Logger, HttpException, HttpStatus } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { IdeaEntity } from '../idea.entity';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -22,15 +22,21 @@ export class IdeaService {
   } 
 
   async readOneIdea(id: string) {
-    return await this.ideaRepository.findOne({where: {id}});
+    const idea =  await this.ideaRepository.findOne({where: {id}});
+    if(!idea) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+    return idea;
   }
 
   async updateIdea(id: string, data: Partial<IdeaDTO>) {
+    const idea = await this.ideaRepository.findOne({where: {id}});
+    if(!idea) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
     await this.ideaRepository.update({id}, data);
-    return await this.ideaRepository.findOne({id});
+    return await this.ideaRepository.findOne({where: {id}});
   }
 
   async deleteIdea(id: string) {
+    const idea = await this.ideaRepository.findOne({where: {id}});
+    if(!idea) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
     await this.ideaRepository.delete({id});
     return {deleted: true};
   }
